+++ scripts/macros/classFeatures/bard/collegeOfSwords/bladeFlourish.js
@@ -1,32 +1,26 @@
+import {constants} from '../../../../constants.js';
 import {chris} from '../../../../helperFunctions.js';
 import {queue} from '../../../../utility/queue.js';
-export async function bladeFlourish({speaker, actor, token, character, item, args, scope, workflow}) {
-    let itemName = workflow.item.name.toLowerCase()
-    if (itemName.includes('booming blade') || itemName.includes('green-flame blade')) return;
-    let sourceActor = workflow.actor;
-    let effect1 = chris.findEffect(sourceActor, 'Blade Flourish Movement');
-    if (workflow.item.type === 'weapon' && !effect1) {
-        let feature0 = sourceActor.items.getName('Blade Flourish Movement');
-        if (feature0) feature0.use();
-    }
-    if (workflow.item.type === 'weapon' && workflow.hitTargets.size === 1) {
-        let effect2 = chris.findEffect(sourceActor, 'Blade Flourish');
-        if (effect2) return;
-        let bardicInspiration = sourceActor.items.getName('Bardic Inspiration');
-        if (!bardicInspiration) bardicInspiration = sourceActor.items.getName('Bardic Inspiration & Magical Inspiration');
-        if (!bardicInspiration) {
-            ui.notifications.warn('Source actor does not appear to have a Bardic Inspiration feature!');
-            return;
-        }
+async function attack({speaker, actor, token, character, item, args, scope, workflow}) {
+    if (workflow.item.flags?.['chris-premades']?.spell?.boomingBlade || workflow.item.flags?.['chris-premades']?.spell?.greenFlameBlade) return;
+    let feature = chris.getItem(workflow.actor, 'Blade Flourish');
+    let feature2 = chris.getItem(workflow.actor, 'Defensive Flourish');
+    let feature3 = chris.getItem(workflow.actor,'Mobile Flourish');
+    let feature4 = chris.getItem(workflow.actor,'Slashing Flourish');
+    let movement = chris.getItem(workflow.actor, 'Blade Flourish Movement');
+    if (!movement || !feature || !feature2 || !feature3 || !feature4) return;
+    let movementEffect = chris.findEffect(workflow.actor, 'Blade Flourish Movement');
+    if (constants.weaponAttacks.includes(workflow.item.system.actionType) && !movementEffect && chris.inCombat()) await movement.use();
+    let perTurnCheck = chris.perTurnCheck(feature, 'feature', 'bladeFlourish', false, workflow.token.id);
+    if (!perTurnCheck) return;
+    if (constants.weaponAttacks.includes(workflow.item.system.actionType) && workflow.hitTargets.size === 1) {
+        let bardicInspiration = chris.getItem(workflow.actor, 'Bardic Inspiration');
+        if (!bardicInspiration) return;
         let queueSetup = await queue.setup(workflow.item.uuid, 'bladeFlourish', 151);
         if (!queueSetup) return;
         let bardicInspirationUses = bardicInspiration.system.uses.value;
-        let classFeature = sourceActor.items.getName('Bard');
-        let skipUses = false;
-        if (classFeature) {
-            let levels = classFeature.system.levels;
-            if (levels >= 14) skipUses = await chris.dialog('Use d6 instead of Bardic Inspiration for Blade Flourish?', [['Yes', true], ['No', false]]);
-        }
+        let mastersFlourish = chris.getItem(workflow.actor, 'Master\'s Flourish');
+        let skipUses = mastersFlourish ? await chris.dialog(mastersFlourish.name, constants.yesNo, 'Use d6 instead of Bardic Inspiration for Blade Flourish?') : false;
         if (bardicInspirationUses <= 0 && !skipUses) {
             queue.remove(workflow.item.uuid);
             return;
@@ -37,44 +31,26 @@ export async function bladeFlourish({speaker, actor, token, character, item, arg
             ['Slashing Flourish', 'SF'],
             ['No', false]
         ];
-        let selectedOption = await chris.dialog('Use a Blade Flourish?', options);
+        let selectedOption = await chris.dialog(feature.name, options, 'Use a Blade Flourish?');
         if (!selectedOption) {
             queue.remove(workflow.item.uuid);
             return;
         }
-        let effectData1 = {
-            'label': 'Blade Flourish',
-            'icon': 'icons/skills/melee/maneuver-sword-katana-yellow.webp',
-            'duration': {'turns': 1},
-            'origin': workflow.item.uuid,
-            'flags': {
-                'dae': {
-                    'specialDuration': [
-                        'combatEnd'
-                    ]
-                }
-            }
-        };
-        await chris.createEffect(sourceActor, effectData1);
+        await chris.setTurnCheck(feature, 'feature', 'bladeFlourish');
         if (!skipUses) bardicInspiration.update({'system.uses.value': bardicInspirationUses - 1});
-        let bardicInspirationDie = sourceActor.system.scale.bard['bardic-inspiration'];
+        let bardicInspirationDie = workflow.actor.system.scale.bard['bardic-inspiration'];
         if (skipUses) bardicInspirationDie = {'formula': '1d6'};
-        if (!bardicInspirationDie) {
-            ui.notifications.warn('Source actor does not appear to have a Bardic Inspiration scale!');
-            queue.remove(workflow.item.uuid);
-            return;
-        }
+        if (!bardicInspirationDie) return;
         let damageType = workflow.item.system.damage.parts[0][1];
         let bonusDamageFormula = bardicInspirationDie.formula + '[' + damageType + ']'
         await chris.addToDamageRoll(workflow, bonusDamageFormula);
         let bardicInspirationDieRoll = workflow.damageRolls[workflow.damageRolls.length - 1].total;
         switch (selectedOption) {
             case 'DF':
-                let feature2 = sourceActor.items.getName('Defensive Flourish');
-                if (feature2) feature2.use();
+                await feature2.use();
                 let effectData2 = {
-                    'label': 'Defensive Flourish',
-                    'icon': 'icons/skills/melee/swords-parry-block-blue.webp',
+                    'label': feature2.name,
+                    'icon': feature2.img,
                     'duration': {'rounds': 1},
                     'changes': [
                         {
@@ -84,7 +60,7 @@ export async function bladeFlourish({speaker, actor, token, character, item, arg
                             'priority': 20
                         }
                     ],
-                    'origin': workflow.item.uuid,
+                    'origin': feature2.uuid,
                     'flags': {
                         'dae': {
                             'specialDuration': [
@@ -93,22 +69,27 @@ export async function bladeFlourish({speaker, actor, token, character, item, arg
                         }
                     }
                 };
-                await chris.createEffect(sourceActor, effectData2);
+                await chris.createEffect(workflow.actor, effectData2);
                 break;
             case 'MF':
-                let feature3 = sourceActor.items.getName('Mobile Flourish');
-                if (feature3) feature3.use();
+                await feature3.use();
+                let distance = Math.floor((5 + bardicInspirationDieRoll) / 5) * 5;
+                let push = await chris.dialog(feature3.name, constants.yesNo, 'Push target ' + (5 + bardicInspirationDieRoll) + ' feet?');
+                if (push) await chris.pushToken(workflow.token, workflow.targets.first(), distance);
                 break;
             case 'SF':
-                let feature4 = await sourceActor.items.getName('Slashing Flourish');
-                if (feature4) feature4.use();
-                let nearbyTargets = chris.findNearby(workflow.token, 5, 'enemy');
-                let hitTokenId = workflow.hitTargets.first().id;
-                let removeIndex = nearbyTargets.findIndex(tok => tok.id === hitTokenId);
-                if (removeIndex != -1) nearbyTargets.splice(removeIndex, 1);
-                await chris.applyDamage([nearbyTargets], bardicInspirationDieRoll, damageType);
+                await feature4.use();
+                let nearbyTargets = chris.findNearby(workflow.token, 5, 'enemy').filter(i => i.document.uuid != workflow.hitTargets.first().document.uuid);
+                await chris.applyWorkflowDamage(workflow.token, await chris.damageRoll(workflow, String(bardicInspirationDieRoll), undefined, true), damageType, nearbyTargets, feature4.name, workflow.itemCardId);
                 break;
         }
         queue.remove(workflow.item.uuid);
     }
+}
+async function combatEnd(parent) {
+    await chris.setTurnCheck(parent, 'feature', 'bladeFlourish', true);
+}
+export let bladeFlourish = {
+    'attack': attack,
+    'combatEnd': combatEnd
 }